Added abcd questions

+++ composeApp/src/commonMain/kotlin/com/group83/a83/model/ABCDModel.kt
@@ -5,5 +5,5 @@ data class ABCDModel(
     val id: Int,
     val question: String,
     val answers: List<AnswerModel>,
-    val correctAnswers: List<Int>,
+    val correctAnswers: List<Long>,
 ) : GameQuestion()
